maj des images
git-svn-id: svn://guim.info/mathieu/public/braldahim/trunk@119 76004128-6536-4c65-8b5e-09d3c7d398dd

+++ fetch.php
@@ -273,7 +273,9 @@ class Fetch {
 				continue;
 			}
 		}
-		file_put_contents('cache/'.date("YmdHi").'-'.uniqid(), $content);
+		if (KEEP_SCRIPT_FILE == "yes") {
+			file_put_contents('cache/'.date("YmdHi").'-'.uniqid(), $content);
+		}
 	}
 	
 	/*
